Splitted kernel32.py
