+++ qiling/os/windows/dlls/__init__.py
@@ -3,7 +3,6 @@
 # Cross Platform and Multi Architecture Advanced Binary Emulation Framework
 # Built on top of Unicorn emulator (www.unicorn-engine.org) 
 from .advapi32 import *
-from .kernel32 import *
 from .msvcrt import *
 from .user32 import *
 from .wininet import *
@@ -12,3 +11,4 @@ from .ntdll import *
 from .ucrtbased import *
 from .wsock32 import *
 from .ole32 import *
+from qiling.os.windows.dlls.kernel32 import *
