+++ qiling/os/windows/dlls/kernel32/__init__.py
@@ -1,0 +1,18 @@
+from .errhandlingapi import *
+from .fibersapi import *
+from .fileapi import *
+from .handleapi import *
+from .heapapi import *
+from .libloaderapi import *
+from .memoryapi import *
+from .processenv import *
+from .profileapi import *
+from .processthreadsapi import *
+from .processenv import *
+from .stringapiset import *
+from .synchapi import *
+from .synchapi import *
+from .winbase import *
+from .winnls import *
+from .winnt import *
+from .wow64apiset import *
