+++ qiling/os/windows/dlls/kernel32/fileapi.py
@@ -1,0 +1,205 @@
+#!/usr/bin/env python3
+#
+# Cross Platform and Multi Architecture Advanced Binary Emulation Framework
+# Built on top of Unicorn emulator (www.unicorn-engine.org)
+
+import struct
+import time
+from qiling.os.windows.const import *
+from qiling.os.fncc import *
+from qiling.os.windows.fncc import *
+from qiling.os.windows.utils import *
+from qiling.os.memory import align
+from qiling.os.windows.thread import *
+from qiling.os.windows.handle import *
+from qiling.exception import *
+
+
+# DWORD GetFileType(
+#   HANDLE hFile
+# );
+@winapi(cc=STDCALL, params={
+    "hFile": HANDLE
+})
+def hook_GetFileType(ql, address, params):
+    hFile = params["hFile"]
+    FILE_TYPE_CHAR = 0x0002
+    if hFile == STD_INPUT_HANDLE or hFile == STD_OUTPUT_HANDLE or hFile == STD_ERROR_HANDLE:
+        ret = FILE_TYPE_CHAR
+    else:
+        raise QlErrorNotImplemented("[!] API not implemented")
+    return ret
+
+
+# HANDLE FindFirstFileA(
+#  LPCSTR             lpFileName,
+#  LPWIN32_FIND_DATAA lpFindFileData
+# );
+@winapi(cc=STDCALL, params={
+    "lpFilename": POINTER,
+    "lpFindFileData": POINTER
+})
+def hook_FindFirstFileA(ql, address, params):
+    pass
+
+
+# HANDLE FindNextFileA(
+#  LPCSTR             lpFileName,
+#  LPWIN32_FIND_DATAA lpFindFileData
+# );
+@winapi(cc=STDCALL, params={
+    "lpFilename": POINTER,
+    "lpFindFileData": POINTER
+})
+def hook_FindNextFileA(ql, address, params):
+    pass
+
+
+# BOOL FindClose(
+#    HANDLE hFindFile
+# );
+@winapi(cc=STDCALL, params={
+    "hFindFile": HANDLE,
+})
+def hook_FindClose(ql, address, params):
+    pass
+
+
+# BOOL ReadFile(
+#   HANDLE       hFile,
+#   LPVOID       lpBuffer,
+#   DWORD        nNumberOfBytesToRead,
+#   LPDWORD      lpNumberOfBytesRead,
+#   LPOVERLAPPED lpOverlapped
+# );
+@winapi(cc=STDCALL, params={
+    "hFile": HANDLE,
+    "lpBuffer": POINTER,
+    "nNumberOfBytesToRead": DWORD,
+    "lpNumberOfBytesRead": POINTER,
+    "lpOverlapped": POINTER
+})
+def hook_ReadFile(ql, address, params):
+    ret = 1
+    hFile = params["hFile"]
+    lpBuffer = params["lpBuffer"]
+    nNumberOfBytesToRead = params["nNumberOfBytesToRead"]
+    lpNumberOfBytesRead = params["lpNumberOfBytesRead"]
+    lpOverlapped = params["lpOverlapped"]
+    if hFile == STD_INPUT_HANDLE:
+        s = ql.stdin.read(nNumberOfBytesToRead)
+        slen = len(s)
+        read_len = slen
+        if slen > nNumberOfBytesToRead:
+            s = s[:nNumberOfBytesToRead]
+            read_len = nNumberOfBytesToRead
+        ql.uc.mem_write(lpBuffer, s)
+        ql.uc.mem_write(lpNumberOfBytesRead, ql.pack(read_len))
+    else:
+        f = ql.handle_manager.get(hFile).file
+        data = f.read(nNumberOfBytesToRead)
+        ql.uc.mem_write(lpBuffer, data)
+        ql.uc.mem_write(lpNumberOfBytesRead, ql.pack32(lpNumberOfBytesRead))
+    return ret
+
+
+# BOOL WriteFile(
+#   HANDLE       hFile,
+#   LPCVOID      lpBuffer,
+#   DWORD        nNumberOfBytesToWrite,
+#   LPDWORD      lpNumberOfBytesWritten,
+#   LPOVERLAPPED lpOverlapped
+# );
+@winapi(cc=STDCALL, params={
+    "hFile": HANDLE,
+    "lpBuffer": POINTER,
+    "nNumberOfBytesToWrite": DWORD,
+    "lpNumberOfBytesWritten": POINTER,
+    "lpOverlapped": POINTER
+})
+def hook_WriteFile(ql, address, params):
+    ret = 1
+    hFile = params["hFile"]
+    lpBuffer = params["lpBuffer"]
+    nNumberOfBytesToWrite = params["nNumberOfBytesToWrite"]
+    lpNumberOfBytesWritten = params["lpNumberOfBytesWritten"]
+    lpOverlapped = params["lpOverlapped"]
+    if hFile == STD_OUTPUT_HANDLE:
+        s = ql.uc.mem_read(lpBuffer, nNumberOfBytesToWrite)
+        ql.stdout.write(s)
+        ql.uc.mem_write(lpNumberOfBytesWritten, ql.pack(nNumberOfBytesToWrite))
+    else:
+        try:
+            f = ql.handle_manager.get(hFile).file
+        except KeyError as ke:
+            # Invalid handle
+            ql.last_error = 0x6  # ERROR_INVALID_HANDLE
+            return 0
+        buffer = ql.uc.mem_read(lpBuffer, nNumberOfBytesToWrite)
+        f.write(bytes(buffer))
+        ql.uc.mem_write(lpNumberOfBytesWritten, ql.pack32(nNumberOfBytesToWrite))
+    return ret
+
+
+def _CreateFile(ql, address, params, name):
+    ret = INVALID_HANDLE_VALUE
+
+    s_lpFileName = params["lpFileName"]
+    dwDesiredAccess = params["dwDesiredAccess"]
+    dwShareMode = params["dwShareMode"]
+    lpSecurityAttributes = params["lpSecurityAttributes"]
+    dwCreationDisposition = params["dwCreationDisposition"]
+    dwFlagsAndAttributes = params["dwFlagsAndAttributes"]
+    hTemplateFile = params["hTemplateFile"]
+
+    # access mask DesiredAccess
+    mode = ""
+    if dwDesiredAccess & GENERIC_WRITE:
+        mode += "wb"
+    else:
+        mode += "r"
+
+    # create thread handle
+    f = open(os.path.join(ql.rootfs, s_lpFileName.replace("\\", os.sep)), mode)
+    new_handle = Handle(file=f)
+    ql.handle_manager.append(new_handle)
+    ret = new_handle.id
+
+    return ret
+
+
+# HANDLE CreateFileA(
+#   LPCSTR                lpFileName,
+#   DWORD                 dwDesiredAccess,
+#   DWORD                 dwShareMode,
+#   LPSECURITY_ATTRIBUTES lpSecurityAttributes,
+#   DWORD                 dwCreationDisposition,
+#   DWORD                 dwFlagsAndAttributes,
+#   HANDLE                hTemplateFile
+# );
+@winapi(cc=STDCALL, params={
+    "lpFileName": STRING,
+    "dwDesiredAccess": DWORD,
+    "dwShareMode": DWORD,
+    "lpSecurityAttributes": POINTER,
+    "dwCreationDisposition": DWORD,
+    "dwFlagsAndAttributes": DWORD,
+    "hTemplateFile": HANDLE
+})
+def hook_CreateFileA(ql, address, params):
+    ret = _CreateFile(ql, address, params, "CreateFileA")
+    return ret
+
+
+@winapi(cc=STDCALL, params={
+    "lpFileName": WSTRING,
+    "dwDesiredAccess": DWORD,
+    "dwShareMode": DWORD,
+    "lpSecurityAttributes": POINTER,
+    "dwCreationDisposition": DWORD,
+    "dwFlagsAndAttributes": DWORD,
+    "hTemplateFile": HANDLE
+})
+def hook_CreateFileW(ql, address, params):
+    ret = _CreateFile(ql, address, params, "CreateFileW")
+    return ret
