+++ qiling/os/windows/dlls/kernel32/profileapi.py
@@ -1,0 +1,26 @@
+#!/usr/bin/env python3
+#
+# Cross Platform and Multi Architecture Advanced Binary Emulation Framework
+# Built on top of Unicorn emulator (www.unicorn-engine.org)
+
+import struct
+import time
+from qiling.os.windows.const import *
+from qiling.os.fncc import *
+from qiling.os.windows.fncc import *
+from qiling.os.windows.utils import *
+from qiling.os.memory import align
+from qiling.os.windows.thread import *
+from qiling.os.windows.handle import *
+from qiling.exception import *
+
+
+# BOOL QueryPerformanceCounter(
+#   LARGE_INTEGER *lpPerformanceCount
+# );
+@winapi(cc=STDCALL, params={
+    "lpPerformanceCount": POINTER
+})
+def hook_QueryPerformanceCounter(ql, address, params):
+    ret = 0
+    return ret
