+++ qiling/os/windows/dlls/kernel32/winbase.py
@@ -1,0 +1,129 @@
+#!/usr/bin/env python3
+#
+# Cross Platform and Multi Architecture Advanced Binary Emulation Framework
+# Built on top of Unicorn emulator (www.unicorn-engine.org)
+
+import struct
+import time
+from qiling.os.windows.const import *
+from qiling.os.fncc import *
+from qiling.os.windows.fncc import *
+from qiling.os.windows.utils import *
+from qiling.os.memory import align
+from qiling.os.windows.thread import *
+from qiling.os.windows.handle import *
+from qiling.exception import *
+
+
+# __analysis_noreturn VOID FatalExit(
+#   int ExitCode
+# );
+@winapi(cc=STDCALL, params={
+    "ExitCode": INT
+})
+def hook_FatalExit(ql, address, params):
+    ql.uc.emu_stop()
+    ql.RUN = False
+
+
+# PVOID EncodePointer(
+#  _In_ PVOID Ptr
+# );
+@winapi(cc=STDCALL, params={
+    "Ptr": POINTER
+})
+def hook_EncodePointer(ql, address, params):
+    return params['Ptr']
+
+
+# PVOID DecodePointer(
+#  _In_ PVOID Ptr
+# );
+@winapi(cc=STDCALL, params={
+    "Ptr": POINTER
+})
+def hook_DecodePointer(ql, address, params):
+    return params['Ptr']
+
+
+# UINT WinExec(
+#   LPCSTR lpCmdLine,
+#   UINT   uCmdShow
+# );
+@winapi(cc=STDCALL, params={
+    "lpCmdLine": STRING,
+    "uCmdShow": UINT
+})
+def hook_WinExec(ql, address, params):
+    return 33
+
+
+# DWORD GetEnvironmentVariableA(
+#   LPCSTR lpName,
+#   LPSTR  lpBuffer,
+#   DWORD  nSize
+# );
+@winapi(cc=STDCALL, params={
+    "lpName": STRING,
+    "lpBuffer": POINTER,
+    "nSize": DWORD
+})
+def hook_GetEnvironmentVariableA(ql, address, params):
+    ret = 0
+    return ret
+
+
+# DECLSPEC_ALLOCATOR HLOCAL LocalAlloc(
+#   UINT   uFlags,
+#   SIZE_T uBytes
+# );
+@winapi(cc=STDCALL, params={
+    "uFlags": UINT,
+    "uBytes": SIZE_T
+})
+def hook_LocalAlloc(ql, address, params):
+    ret = ql.heap.mem_alloc(params["uBytes"])
+    return ret
+
+
+# UINT SetHandleCount(
+#   UINT uNumber
+# );
+@winapi(cc=STDCALL, params={
+    "uNumber": UINT
+})
+def hook_SetHandleCount(ql, address, params):
+    uNumber = params["uNumber"]
+    return uNumber
+
+
+# LPVOID GlobalLock(
+#  HGLOBAL hMem
+# );
+@winapi(cc=STDCALL, params={
+    "hMem": POINTER
+})
+def hook_GlobalLock(ql, address, params):
+    return params['hMem']
+
+
+# LPVOID GlobalUnlock(
+#  HGLOBAL hMem
+# );
+@winapi(cc=STDCALL, params={
+    "hMem": POINTER
+})
+def hook_GlobalUnlock(ql, address, params):
+    return 1
+
+
+# DECLSPEC_ALLOCATOR HGLOBAL GlobalAlloc(
+#  UINT   uFlags,
+#  SIZE_T dwBytes
+# );
+@winapi(cc=STDCALL, params={
+    "uFlags": UINT,
+    "dwBytes": UINT
+})
+def hook_GlobalAlloc(ql, address, params):
+    return ql.heap.mem_alloc(params['dwBytes'])
