+++ qiling/os/windows/dlls/kernel32/winnls.py
@@ -1,0 +1,161 @@
+#!/usr/bin/env python3
+#
+# Cross Platform and Multi Architecture Advanced Binary Emulation Framework
+# Built on top of Unicorn emulator (www.unicorn-engine.org)
+
+import struct
+import time
+from qiling.os.windows.const import *
+from qiling.os.fncc import *
+from qiling.os.windows.fncc import *
+from qiling.os.windows.utils import *
+from qiling.os.memory import align
+from qiling.os.windows.thread import *
+from qiling.os.windows.handle import *
+from qiling.exception import *
+
+
+# BOOL SetThreadLocale(
+#   LCID Locale
+# );
+@winapi(cc=STDCALL, params={
+    "Locale": UINT
+})
+def hook_SetThreadLocale(ql, address, params):
+    return 0xC000  # LOCALE_CUSTOM_DEFAULT
+
+
+# UINT GetACP(
+# );
+@winapi(cc=STDCALL, params={})
+def hook_GetACP(ql, address, params):
+    ret = 437
+    return ret
+
+
+# BOOL GetCPInfo(
+#   UINT     CodePage,
+#   LPCPINFO lpCPInfo
+# );
+@winapi(cc=STDCALL, params={
+    "CodePage": UINT,
+    "lpCPInfo": POINTER
+})
+def hook_GetCPInfo(ql, address, params):
+    ret = 1
+    return ret
+
+
+# int GetLocaleInfoA(
+#   LCID   Locale,
+#   LCTYPE LCType,
+#   LPSTR  lpLCData,
+#   int    cchData
+# );
+@winapi(cc=STDCALL, params={
+    "Locale": DWORD,
+    "LCType": DWORD,
+    "lpLCData": POINTER,
+    "cchData": INT,
+})
+def hook_GetLocaleInfoA(ql, address, params):
+    locale_value = params["Locale"]
+    lctype_value = params["LCType"]
+    cchData = params["cchData"]
+    ql.dprint("locale")
+    ql.dprint(locale_value)
+    ql.dprint("lctype")
+    ql.dprint(lctype_value)
+
+    local_dict = LOCALE.get(locale_value, None)
+    if local_dict is None:
+        raise QlErrorNotImplemented("[!] API not implemented")
+    lctype = local_dict[lctype_value] + "\x00"
+
+    if cchData != 0:
+        lplcdata = params["lpLCData"]
+        ql.uc.mem_write(lplcdata, lctype.encode("utf16-le"))
+    return len(lctype)
+
+
+# BOOL IsValidCodePage(
+#  UINT CodePage
+# );
+@winapi(cc=STDCALL, params={
+    "CodePage": UINT
+})
+def hook_IsValidCodePage(ql, address, params):
+    return 1
+
+
+# int LCMapStringW(
+#   LCID    Locale,
+#   DWORD   dwMapFlags,
+#   LPCWSTR lpSrcStr,
+#   int     cchSrc,
+#   LPWSTR  lpDestStr,
+#   int     cchDest
+# );
+@winapi(cc=STDCALL, params={
+    "Locale": POINTER,
+    "dwMapFlags": DWORD,
+    "lpSrcStr": WSTRING,
+    "cchSrc": INT,
+    "lpDestStr": POINTER,
+    "cchDest": INT
+})
+def hook_LCMapStringW(ql, address, params):
+    # TODO implement
+    ret = 1
+    return ret
+
+
+# int LCMapStringA(
+#   LCID   Locale,
+#   DWORD  dwMapFlags,
+#   LPCSTR lpSrcStr,
+#   int    cchSrc,
+#   LPSTR  lpDestStr,
+#   int    cchDest
+# );
+@winapi(cc=STDCALL, params={
+    "Locale": POINTER,
+    "dwMapFlags": DWORD,
+    "lpSrcStr": WSTRING,
+    "cchSrc": INT,
+    "lpDestStr": POINTER,
+    "cchDest": INT
+})
+def hook_LCMapStringA(ql, address, params):
+    # TODO implement
+    ret = 1
+    return ret
+
+
+# int LCMapStringEx(
+#   LPCWSTR          lpLocaleName,
+#   DWORD            dwMapFlags,
+#   LPCWSTR          lpSrcStr,
+#   int              cchSrc,
+#   LPWSTR           lpDestStr,
+#   int              cchDest,
+#   LPNLSVERSIONINFO lpVersionInformation,
+#   LPVOID           lpReserved,
+#   LPARAM           sortHandle
+# );
+@winapi(cc=STDCALL, params={
+    "lpLocaleName": POINTER,
+    "dwMapFlags": DWORD,
+    "lpSrcStr": POINTER,
+    "cchSrc": INT,
+    "lpDestStr": POINTER,
+    "cchDest": INT,
+    "lpVersionInformation": POINTER,
+    "lpReserved": UINT,
+    "sortHandle": UINT
+
+})
+def hook_LCMapStringEx(ql, address, params):
+    # TODO needs a better implementation
+    ret = 1
+    return ret
