+++ qiling/os/windows/dlls/kernel32/wow64apiset.py
@@ -1,0 +1,34 @@
+#!/usr/bin/env python3
+#
+# Cross Platform and Multi Architecture Advanced Binary Emulation Framework
+# Built on top of Unicorn emulator (www.unicorn-engine.org)
+
+import struct
+import time
+from qiling.os.windows.const import *
+from qiling.os.fncc import *
+from qiling.os.windows.fncc import *
+from qiling.os.windows.utils import *
+from qiling.os.memory import align
+from qiling.os.windows.thread import *
+from qiling.os.windows.handle import *
+from qiling.exception import *
+
+
+# BOOL IsWow64Process(
+#   HANDLE hProcess,
+#   PBOOL  Wow64Process
+# );
+@winapi(cc=STDCALL, params={
+    "hProcess": HANDLE,
+    "Wow64Process": POINTER
+})
+def hook_IsWow64Process(ql, address, params):
+    pointer = params["Wow64Process"]
+    false = 0x0.to_bytes(length=ql.pointersize, byteorder='little')
+    true = 0x1.to_bytes(length=ql.pointersize, byteorder='little')
+    if ql.archbit == 32:
+        ql.uc.mem_write(pointer, false)
+    else:
+        raise QlErrorNotImplemented("[!] API not implemented")
+    return 1
