+++ qiling/os/windows/x8664.py
@@ -8,7 +8,7 @@ import types
 from unicorn import *
 from unicorn.x86_const import *
 
-# impport read_string and other commom utils.
+# import read_string and other common utils.
 from qiling.loader.pe import PE, Shellcode
 from qiling.os.windows.dlls import *
 from qiling.os.utils import *
