+++ tests/test_pe_nu.py
@@ -5,23 +5,28 @@
 
 import sys
 
-sys.path.insert(0, "..")
 from qiling import *
 from qiling.exception import *
 from qiling.os.windows.fncc import *
 from qiling.os.windows.utils import *
+sys.path.insert(0, "..")
 
-X86_WIN = unhexlify('fce8820000006089e531c0648b50308b520c8b52148b72280fb74a2631ffac3c617c022c20c1cf0d01c7e2f252578b52108b4a3c8b4c1178e34801d1518b592001d38b4918e33a498b348b01d631ffacc1cf0d01c738e075f6037df83b7d2475e4588b582401d3668b0c4b8b581c01d38b048b01d0894424245b5b61595a51ffe05f5f5a8b12eb8d5d6a01eb2668318b6f87ffd5bbf0b5a25668a695bd9dffd53c067c0a80fbe07505bb4713726f6a0053ffd5e8d5ffffff63616c6300')
-X8664_WIN = unhexlify('fc4881e4f0ffffffe8d0000000415141505251564831d265488b52603e488b52183e488b52203e488b72503e480fb74a4a4d31c94831c0ac3c617c022c2041c1c90d4101c1e2ed5241513e488b52203e8b423c4801d03e8b80880000004885c0746f4801d0503e8b48183e448b40204901d0e35c48ffc93e418b34884801d64d31c94831c0ac41c1c90d4101c138e075f13e4c034c24084539d175d6583e448b40244901d0663e418b0c483e448b401c4901d03e418b04884801d0415841585e595a41584159415a4883ec204152ffe05841595a3e488b12e949ffffff5d49c7c1000000003e488d95fe0000003e4c8d850f0100004831c941ba45835607ffd54831c941baf0b5a256ffd548656c6c6f2c2066726f6d204d534621004d657373616765426f7800') 
+X86_WIN = unhexlify(
+    'fce8820000006089e531c0648b50308b520c8b52148b72280fb74a2631ffac3c617c022c20c1cf0d01c7e2f252578b52108b4a3c8b4c1178e34801d1518b592001d38b4918e33a498b348b01d631ffacc1cf0d01c738e075f6037df83b7d2475e4588b582401d3668b0c4b8b581c01d38b048b01d0894424245b5b61595a51ffe05f5f5a8b12eb8d5d6a01eb2668318b6f87ffd5bbf0b5a25668a695bd9dffd53c067c0a80fbe07505bb4713726f6a0053ffd5e8d5ffffff63616c6300')
+X8664_WIN = unhexlify(
+    'fc4881e4f0ffffffe8d0000000415141505251564831d265488b52603e488b52183e488b52203e488b72503e480fb74a4a4d31c94831c0ac3c617c022c2041c1c90d4101c1e2ed5241513e488b52203e8b423c4801d03e8b80880000004885c0746f4801d0503e8b48183e448b40204901d0e35c48ffc93e418b34884801d64d31c94831c0ac41c1c90d4101c138e075f13e4c034c24084539d175d6583e448b40244901d0663e418b0c483e448b401c4901d03e418b04884801d0415841585e595a41584159415a4883ec204152ffe05841595a3e488b12e949ffffff5d49c7c1000000003e488d95fe0000003e4c8d850f0100004831c941ba45835607ffd54831c941baf0b5a256ffd548656c6c6f2c2066726f6d204d534621004d657373616765426f7800')
+
 
 def test_pe_win_x8664_hello():
-    ql = Qiling(["../examples/rootfs/x8664_windows/bin/x8664_hello.exe"], "../examples/rootfs/x8664_windows", output = "default")
+    ql = Qiling(["../examples/rootfs/x8664_windows/bin/x8664_hello.exe"], "../examples/rootfs/x8664_windows",
+                output="default")
     ql.run()
     del ql
 
 
 def test_pe_win_x86_hello():
-    ql = Qiling(["../examples/rootfs/x86_windows/bin/x86_hello.exe"], "../examples/rootfs/x86_windows", output = "default", log_dir='.', log_split=True)
+    ql = Qiling(["../examples/rootfs/x86_windows/bin/x86_hello.exe"], "../examples/rootfs/x86_windows",
+                output="default", log_dir='.', log_split=True)
     ql.run()
     del ql
 
@@ -57,7 +62,7 @@ def test_pe_win_x86_regdemo():
 
 
 def test_pe_win_x8664_fls():
-    ql = Qiling(["../examples/rootfs/x8664_windows/bin/Fls.exe"], "../examples/rootfs/x8664_windows", output = "default")
+    ql = Qiling(["../examples/rootfs/x8664_windows/bin/Fls.exe"], "../examples/rootfs/x8664_windows", output="default")
     ql.run()
     del ql
 
@@ -66,6 +71,7 @@ def test_pe_win_x86_wannacry():
     def stop(ql):
         print("killerswtichfound")
         ql.uc.emu_stop()
+
     ql = Qiling(["../examples/rootfs/x86_windows/bin/wannacry.bin"], "../examples/rootfs/x86_windows")
     ql.hook_address(stop, 0x40819a)
     ql.run()
@@ -73,13 +79,15 @@ def test_pe_win_x86_wannacry():
 
 
 def test_windowssc_x86():
-    ql = Qiling(shellcoder = X86_WIN, archtype = "x86", ostype = "windows", rootfs="../examples/rootfs/x86_windows", output="default")
+    ql = Qiling(shellcoder=X86_WIN, archtype="x86", ostype="windows", rootfs="../examples/rootfs/x86_windows",
+                output="default")
     ql.run()
     del ql
 
 
 def test_windowssc_x64():
-    ql = Qiling(shellcoder = X8664_WIN, archtype = "x8664", ostype = "windows", rootfs="../examples/rootfs/x8664_windows", output="debug")
+    ql = Qiling(shellcoder=X8664_WIN, archtype="x8664", ostype="windows", rootfs="../examples/rootfs/x8664_windows",
+                output="debug")
     ql.run()
     del ql
 
@@ -95,14 +103,14 @@ def test_pe_win_x8664_customapi():
         ret = len(string)
         return ret
 
-
     def my_sandbox(path, rootfs):
-        ql = Qiling(path, rootfs, output = "debug")
+        ql = Qiling(path, rootfs, output="debug")
         ql.set_syscall("puts", my_puts64)
         ql.run()
         del ql
 
     my_sandbox(["../examples/rootfs/x8664_windows/bin/x8664_hello.exe"], "../examples/rootfs/x8664_windows")
+
 
 def test_pe_win_x86_crackme():
     class StringBuffer:
@@ -114,7 +122,7 @@ def test_pe_win_x86_crackme():
             self.buffer = self.buffer[n:]
             return ret
 
-        def readline(self, end = b'\n'):
+        def readline(self, end=b'\n'):
             ret = b''
             while True:
                 c = self.read(1)
@@ -127,7 +135,6 @@ def test_pe_win_x86_crackme():
             self.buffer += string
             return len(string)
 
-
     def force_call_dialog_func(ql):
         # get DialogFunc address
         lpDialogFunc = ql.unpack32(ql.mem_read(ql.sp - 0x8, 4))
@@ -139,7 +146,6 @@ def test_pe_win_x86_crackme():
         ql.stack_push(0x0401018)
         # force EIP to DialogFunc
         ql.pc = lpDialogFunc
-
 
     def our_sandbox(path, rootfs):
         ql = Qiling(path, rootfs)
@@ -154,6 +160,7 @@ def test_pe_win_x86_crackme():
         del ql
 
     our_sandbox(["../examples/rootfs/x86_windows/bin/Easy_CrackMe.exe"], "../examples/rootfs/x86_windows")
+
 
 def test_pe_win_x86_bruteforcecrackme():
     class StringBuffer:
@@ -174,12 +181,12 @@ def test_pe_win_x86_bruteforcecrackme():
             self.buffer += string
             return len(string)
 
-
     def instruction_count(ql, address, size, user_data):
         user_data[0] += 1
 
     def get_count(flag):
-        ql = Qiling(["../examples/rootfs/x86_windows/bin/crackme.exe"], "../examples/rootfs/x86_windows", libcache = True, output = "off")
+        ql = Qiling(["../examples/rootfs/x86_windows/bin/crackme.exe"], "../examples/rootfs/x86_windows", libcache=True,
+                    output="off")
         ql.stdin = StringBuffer()
         ql.stdout = StringBuffer()
         ql.stdin.write(bytes("".join(flag) + "\n", 'utf-8'))
@@ -190,11 +197,10 @@ def test_pe_win_x86_bruteforcecrackme():
         ql.nprint(" ============ count: %d ============ " % count[0])
         return count[0]
 
-
     def solve():
         # BJWXB_CTF{C5307D46-E70E-4038-B6F9-8C3F698B7C53}
         prefix = list("BJWXB_CTF{C5307D46-E70E-4038-B6F9-8C3F698B7C")
-        flag = list("\x00"*100)
+        flag = list("\x00" * 100)
         base = get_count(prefix + flag)
         i = 0
 
@@ -213,7 +219,8 @@ def test_pe_win_x86_bruteforcecrackme():
         except KeyboardInterrupt:
             print("STOP: KeyboardInterrupt")
 
-    solve()        
+    solve()
+
 
 if __name__ == "__main__":
     test_pe_win_x8664_hello()
@@ -227,6 +234,6 @@ if __name__ == "__main__":
     test_pe_win_x86_regdemo()
     test_pe_win_x86_crackme()
     test_pe_win_x8664_customapi()
-    #test_windowssc_x86()
-    #test_windowssc_x64()
-    #test_pe_win_x86_bruteforcecrackme()
+    # test_windowssc_x86()
+    # test_windowssc_x64()
+    # test_pe_win_x86_bruteforcecrackme()
